Novas práticas de deleção e refatoração do future/grid do notesMenu para acabar com flickering

+++ lib/components/personListTile.dart
@@ -29,6 +29,10 @@ class Personlisttile extends StatelessWidget {
           child: Row(
             mainAxisAlignment: MainAxisAlignment.start,
             children: [
+              Padding(
+                padding: const EdgeInsets.only(right: 8.0),
+                child: CircleAvatar(),
+              ),
               Column(
                 crossAxisAlignment: CrossAxisAlignment.start,
                 children: [
@@ -50,10 +54,44 @@ class Personlisttile extends StatelessWidget {
               ),
               IconButton(
                 onPressed: () {
-                  Provider.of<Personservice>(
-                    context,
-                    listen: false,
-                  ).deletePerson(person!);
+                  showDialog(
+                    context: context,
+                    builder: (context) {
+                      return AlertDialog(
+                        title: const Text(
+                          "Deseja realmente deletar essa pessoa da lista?",
+                          style: TextStyle(
+                            fontWeight: FontWeight.bold,
+                            fontSize: 20,
+                          ),
+                        ),
+                        content: const Text(
+                          'Todas as notas referente a essa pessoa também serão excluídas permanentemente',
+                        ),
+                        actions: [
+                          TextButton(
+                            onPressed: () {
+                              Navigator.of(context).pop();
+                            },
+                            child: Text("Cancelar"),
+                          ),
+                          TextButton(
+                            onPressed: () async {
+                              await Provider.of<Personservice>(
+                                context,
+                                listen: false,
+                              ).deletePerson(person!);
+                              Navigator.of(context).pop();
+                            },
+                            child: Text(
+                              "Excluir",
+                              style: TextStyle(color: Colors.red),
+                            ),
+                          ),
+                        ],
+                      );
+                    },
+                  );
                 },
                 icon: Icon(Icons.delete, color: Colors.red),
               ),
